requirements.txt
curl_cffi>=0.10.0
requests>=2.31.0 
bs4>=0.0.1
python-dotenv>=1.0.1
schedule>=1.2.0